Phase-III__Todo-AI-Chatbot/backend/pyproject.toml
[tool.poetry]
name = "ai-todo-chatbot-backend"
version = "1.0.0"
description = "AI-Powered Todo Chatbot Backend using FastAPI, OpenAI Agents SDK with Gemini LLM"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
openai = "^1.58.0"
openai-agents = "^0.0.8"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.14.0"
tenacity = "^9.0.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
sse-starlette = "^2.1.0"
mcp = "^1.0.0"
httpx = "^0.28.0"
greenlet = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
